being more explicit about h3
After looking at h3 I can definitely see the unmatched performance as its dep free etc. But when we say deployment versatility. 
is It because of `Portable: Works perfectly in Serverless, Workers, and Node.js ?`

If that is the case in terms of reading and even SEO. Wouldn't be better to say something like my update suggestion

I know it's picky and it's written in h3 doc but may be worth to add it inside nuxt concept doc what do you think ?