Add examples of multi filed components
For our design system we use a folder for every component (which sometimes contains other component parts). To get this to work in our nuxt.js export we had to create an array of directories and loop over it to add them all to the export. I thought maybe others would be helped with it :) We also didn't want to use a prefix so I added an example for that as well. I also changed the import order, I got an error if I didn't import pathe after defineNuxtModule so I switched them around. I used @nuxt/kit-edge in this example, as I got an error when I used @nuxt/kit.