feat(nuxi): implement missing listen hook
The listen hook is useful to expose the address that the local development server is listening on, some third party modules will depend on fetching this reliably.

As far as I can tell the listen hook is unused in nuxt 3 so far, this types seem to be consistent with what was there, I've just typed it.