Added a "Are you courageously Nuxt" subsection
The github issues seem to be hovering at 230 or so, and when one gets resolved, another few open up. It may be a good idea to motivate readers to take a stab at contributing, and perhaps consider a "Are You Nuxt?" campaign. If enough people contribute, the burn rate should improve greatly. If the merch were designed cool enough (hats / shirts /hoodies) it may motivate people.