Added data fetching migration
This is a start to all of the different options provided by Nuxt 3 for data fetching. I wrote this up to where I was with the migration of our app. I know there is a ton more that we will have to touch on.

@danielroe : The big question I came across when migrating our app from Nuxt 2 to Nuxt 3 was how to design reusable functions that call an API through fetch. For example, if I were to make an API endpoint that accepted a `search` term or a `pagination` term, what would be the best way to abstract that? Interface with the provided `$fetch` object?

I'll keep updating this as I migrate pieces of our app as well. If there's any changes, let me know!