Add frequently asked question
After the `build.transpile` is added, there is a problem that `[nuxt] [request error] [unhandled] [500] require is not defined in ES module scope, you can use import instead`. You can use import instead. Most people do not seem to know how to solve it, so it is suggested to add solutions to reduce `issue`