docs: Add recommendation for controlling plugin registration order.
Controlling the order of plugin registration is a common requirement, so I think this should be added as an explicit recommendation in the docs.

Right now, it just says "Plugins are called in order sequentially and before everything else" in a small part of the docs. I don't think it's verbose enough. A section titled something like "Plugin registration order" would be very helpful.

From discussion https://github.com/nuxt/framework/discussions/1260#discussioncomment-3842023